Bind opening the menu on tap
An attempt to fix this issue:
https://github.com/JedWatson/react-select/pull/245
This performs quite nicely however it causes the dropdown menu to
flicker off and on when the control is already open when clicked on desktop,
on mobile there are no issues. I've looked into fixing this however I think
it requires a fair amount of tidying up of how focus is applied to the input
sitting behind the select menu. Issuing a PR in the meantime if anyone else
has ideas or guidence